+++ CHANGELOG.md
@@ -17,6 +17,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
+
+## [0.4.2]
+
+### Added
+- `automargin` property to `Axis` according to https://plotly.com/python/reference/layout/xaxis/#layout-xaxis-automargin
+
+### Fixed
+- Remove unnecessary `kotlinx-css` dependency.
+- Added compatibility mode for legacy notebooks. Use `Plotly.jupyter.notebook()` call to enable legacy mode.
+
+### Security
 ## [0.4.0]
 ### Added
 - Jupyter integration plugin for server